+++ libs/platform/src/lib/components/form/combobox/commons/base-combobox.ts
@@ -40,7 +40,8 @@ import {
     ListComponent,
     MobileModeConfig,
     TemplateDirective,
-    FormStates
+    FormStates,
+    closestElement
 } from '@fundamental-ngx/core';
 import {
     ArrayComboBoxDataSource,
@@ -271,8 +272,15 @@ export abstract class BaseCombobox extends CollectionBaseInput implements AfterV
     private _element: HTMLElement = this.elementRef.nativeElement;
     /** Keys, that won't trigger the popover's open state, when dispatched on search input */
     private readonly _nonOpeningKeys: number[] = [
-        ESCAPE, ENTER, CONTROL, TAB, SHIFT,
-        UP_ARROW, RIGHT_ARROW, DOWN_ARROW, LEFT_ARROW
+        ESCAPE,
+        ENTER,
+        CONTROL,
+        TAB,
+        SHIFT,
+        UP_ARROW,
+        RIGHT_ARROW,
+        DOWN_ARROW,
+        LEFT_ARROW
     ];
 
     /** @hidden */
@@ -365,20 +373,18 @@ export abstract class BaseCombobox extends CollectionBaseInput implements AfterV
     /** @hidden
      * Close list
      * */
-    close(event: MouseEvent = null, forceClose: boolean = false): void {
-        if (event) {
-            const target = event.target as HTMLInputElement;
-            if (target && target.id === this.id) {
-                return;
-            }
+    close(event: MouseEvent): void {
+        const target = event.target as HTMLInputElement;
+        if (target && target.id === this.id) {
+            return;
         }
 
-        if (this.isOpen && (forceClose || this.canClose)) {
-            this.isOpen = false;
-            this.openChange.next(this.isOpen);
-            this.cd.markForCheck();
-            this.onTouched();
+        if (!!closestElement(`#${this.id}-input-group-container`, event.target)) {
+            event.preventDefault();
+            event.stopPropagation();
         }
+
+        this.showList(false);
     }
 
     /** @hidden */
@@ -432,11 +438,8 @@ export abstract class BaseCombobox extends CollectionBaseInput implements AfterV
 
         this.showList(!isOpen);
 
-        if (!this.mobile) {
-            this.searchInputElement.nativeElement.focus();
-        }
-        if (this.isOpen) {
-            this.listComponent.setItemActive(0);
+        if (this.isOpen) {	
+            this.listComponent.setItemActive(0);	
         }
     }
 
@@ -471,9 +474,10 @@ export abstract class BaseCombobox extends CollectionBaseInput implements AfterV
             this.showList(false);
         } else if (!event.ctrlKey && !KeyUtil.isKeyCode(event, this._nonOpeningKeys)) {
             this.showList(true);
-            const acceptedKeys = !KeyUtil.isKeyCode(event, BACKSPACE)
-                && !KeyUtil.isKeyType(event, 'alphabetical')
-                && !KeyUtil.isKeyType(event, 'numeric');
+            const acceptedKeys =
+                !KeyUtil.isKeyCode(event, BACKSPACE) &&
+                !KeyUtil.isKeyType(event, 'alphabetical') &&
+                !KeyUtil.isKeyType(event, 'numeric');
             if (this.isEmptyValue && acceptedKeys) {
                 this.listComponent?.setItemActive(0);
             }
@@ -489,7 +493,7 @@ export abstract class BaseCombobox extends CollectionBaseInput implements AfterV
 
     /** @hidden */
     protected get ds(): ComboBoxDataSource<any> {
-        return (<ComboBoxDataSource<any>>this.dataSource);
+        return <ComboBoxDataSource<any>>this.dataSource;
     }
 
     /** @hidden
@@ -586,10 +590,6 @@ export abstract class BaseCombobox extends CollectionBaseInput implements AfterV
     private _initWindowResize(): void {
         this._getOptionsListWidth();
 
-        if (!this.autoResize) {
-            return;
-        }
-
         fromEvent(window, 'resize')
             .pipe(takeUntil(this._destroyed))
             .subscribe(() => this._getOptionsListWidth());
@@ -598,11 +598,11 @@ export abstract class BaseCombobox extends CollectionBaseInput implements AfterV
     /** @hidden */
     private _getOptionsListWidth(): void {
         const body = document.body;
-        const rect = (this._element.querySelector('fd-input-group') as HTMLElement)
-            .getBoundingClientRect();
+        const rect = (this._element.querySelector('fd-input-group') as HTMLElement).getBoundingClientRect();
         const scrollBarWidth = body.offsetWidth - body.clientWidth;
-        this.maxWidth = (window.innerWidth - scrollBarWidth) - rect.left;
+        this.maxWidth = this.autoResize ? window.innerWidth - scrollBarWidth - rect.left : this.minWidth;
         this.minWidth = rect.width - 2;
+        this._cd.detectChanges();
     }
 
     /**